chore: seperated type decleration into own file
