+++ src/core_modules/capture-core/actions/navigateToEnrollmentOverview/navigateToEnrollmentOverview.types.js
@@ -1,0 +1,8 @@
+// @flow
+
+export type NavigateToEnrollmentOverviewProps = {|
+    teiId?: string,
+    programId?: string,
+    orgUnitId?: string,
+    enrollmentId?: string,
+|}
