+++ src/core_modules/capture-core/components/DataStore/DataStore.types.js
@@ -1,0 +1,9 @@
+// @flow
+
+export const actionTypes = {
+    FETCH_DATA_STORE: 'useNewDashboard.FetchDataStore',
+};
+
+export type UseNewDashboard = {|
+    [key: string]: string,
+|}
